Mod : Item 이미지를 string -> url로 변경
