+++ src/main/resources/application-secret.yml
@@ -20,3 +20,7 @@ cloud:
     region:
       static: ap-northeast-2
 
+url:
+  item: https://apoorapoors3.s3.ap-northeast-2.amazonaws.com/poor/
+
+
